A bunch of changes:
Moved a lot of Ubuntu specific preflight/postinstall cruft to the distro plugin where it belongs. Stuff like the EC2_init installation, adding the PPA, installing a specific sudoers file... That's all been moved to ubuntu/distro.py.

As a consequence, I've moved over 3 templates from ec2 to ubuntu, in an ec2 subdirectory so that they don't conflict with the rest of ubuntu's templates. (EC2 is supposed to install a different sudoers, for example)

What remains for EC2's preflight/postinstall is obvious sanity checks, disabling the hwclock, and installing /etc/ec2_version in the AMI.

Renamed option --no-amazon-tools to --ec2-no-amazon-tools (and same for no-euca-tools) for consistency.

VMbuilder EC2 no longer fatally crashes if you don't have the proper tools to deploy an image if you aren't trying to deploy an image. It will, however, warn you that you should have them installed. As far as I know you only need them for bundling or uploading, and we shouldn't crash if we aren't using the tools.

TODO: Write Debian preflight/postinstall for EC2.